#245 Move table and job managing methods out of persistence classes.
